Fixed UVIndex animation
